Создан рабочий контроллер
